+++ spotty/deployment/abstract_instance_volume.py
@@ -1,0 +1,12 @@
+from abc import ABC
+
+
+class AbstractInstanceVolume(ABC):
+
+    @property
+    def name(self) -> str:
+        raise NotImplementedError
+
+    @property
+    def mount_dir(self) -> str:
+        raise NotImplementedError
