+++ spotty/providers/aws/aws_resources/image.py
