+++ spotty/providers/aws/aws_resources/instance.py
