+++ spotty/providers/aws/aws_resources/snapshot.py
