+++ spotty/providers/aws/aws_resources/stack.py
