+++ spotty/providers/aws/deployment/cf_templates/data/create_ami.yaml
