+++ spotty/providers/aws/deployment/cf_templates/data/create_instance_profile.yaml
