+++ spotty/providers/aws/deployment/cf_templates/data/run_container.yaml
