+++ spotty/providers/aws/deployment/cf_templates/instance_profile_template.py
@@ -1,0 +1,8 @@
+import os
+
+
+def prepare_instance_profile_template():
+    with open(os.path.join(os.path.dirname(__file__), 'data', 'create_instance_profile.yaml')) as f:
+        template = f.read()
+
+    return template
