+++ spotty/providers/aws/deployment/project_resources/bucket.py
