+++ spotty/providers/aws/deployment/project_resources/key_pair.py
